+++ components/sortable/SortableReference.tsx
@@ -1,6 +1,10 @@
 "use client";
 import React, { Dispatch, useState } from "react";
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@/components/ui/popover";
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 import { TrashIcon } from "lucide-react";
@@ -13,6 +17,17 @@ import {
   AccordionItem,
   AccordionTrigger,
 } from "@/components/ui/accordion";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { format } from "date-fns";
 import { Textarea } from "@/components/ui/textarea";
 import { useReference } from "@/statemanagement/useReference";
@@ -78,7 +93,7 @@ const SortableReference: React.FC<SortableReferenceProps> = ({
                 <div className="w-full flex justify-start items-center gap-8">
                   <div className="w-1/2 space-y-2">
                     <Label className="capitalize font-normal text-sm text-charcoal flex gap-2 justify-start items-center">
-                      referent's full name
+                      referent&apos;s full name
                     </Label>
                     <Input
                       value={referenceHistory[index]?.referenceFullName || ""}
@@ -130,17 +145,42 @@ const SortableReference: React.FC<SortableReferenceProps> = ({
               </AccordionContent>
             </AccordionItem>
           </Accordion>
-          <div
-            onClick={() => {
-              handleDeleteDiv(index);
-              setReferenceCompany(index, "");
-              setReferenceEmail(index, "");
-              setReferenceFullName(index, "");
-              setReferencePhone(index, "");
-            }}
-          >
-            <TrashIcon className="hover:text-aquamarine-100" />
-          </div>
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <div>
+                <TrashIcon className="hover:text-aquamarine-100" />
+              </div>
+            </AlertDialogTrigger>
+
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle className="text-2xl">
+                  Delete Entry
+                </AlertDialogTitle>
+                <AlertDialogDescription className="text-base text-charcoal">
+                  Are you sure want to delete this entry ?
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+
+              <AlertDialogFooter>
+                <AlertDialogAction
+                  className="bg-aquamarine-100 hover:bg-aquamarine-200 text-white hover:text-white uppercase text-base font-light min-w-[91.5px]"
+                  onClick={() => {
+                    handleDeleteDiv(index);
+                    setReferenceCompany(index, "");
+                    setReferenceEmail(index, "");
+                    setReferenceFullName(index, "");
+                    setReferencePhone(index, "");
+                  }}
+                >
+                  delete
+                </AlertDialogAction>
+                <AlertDialogCancel className="text-black bg-transparent border-charcoal uppercase text-base font-light min-w-[91.5px]">
+                  cancel
+                </AlertDialogCancel>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </div>
       ))}
     </>
