+++ components/sortable/SortableSkills.tsx
@@ -12,6 +12,17 @@ import {
   AccordionItem,
   AccordionTrigger,
 } from "@/components/ui/accordion";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { useSkills } from "@/statemanagement/useSkills";
 import SkillsBadge from "../site/home/SkillsBadge";
 
@@ -174,15 +185,40 @@ const SortableSkills: React.FC<SortableSkillsProps> = ({
               </AccordionContent>
             </AccordionItem>
           </Accordion>
-          <div
-            onClick={() => {
-              handleDeleteDiv(index);
-              setSkillsTitle(index, "");
-              setSkillsLevel(index, "");
-            }}
-          >
-            <TrashIcon className="hover:text-aquamarine-100" />
-          </div>
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <div>
+                <TrashIcon className="hover:text-aquamarine-100" />
+              </div>
+            </AlertDialogTrigger>
+
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle className="text-2xl">
+                  Delete Entry
+                </AlertDialogTitle>
+                <AlertDialogDescription className="text-base text-charcoal">
+                  Are you sure want to delete this entry ?
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+
+              <AlertDialogFooter>
+                <AlertDialogAction
+                  className="bg-aquamarine-100 hover:bg-aquamarine-200 text-white hover:text-white uppercase text-base font-light min-w-[91.5px]"
+                  onClick={() => {
+                    handleDeleteDiv(index);
+                    setSkillsTitle(index, "");
+                    setSkillsLevel(index, "");
+                  }}
+                >
+                  delete
+                </AlertDialogAction>
+                <AlertDialogCancel className="text-black bg-transparent border-charcoal uppercase text-base font-light min-w-[91.5px]">
+                  cancel
+                </AlertDialogCancel>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </div>
       ))}
     </>
